+++ src/js/App/QuickStart/QuickStartCatalog.js
@@ -1,4 +1,6 @@
 import React from 'react';
 import { QuickStartCatalogPage as PfQuickStartCatalogPage } from '@patternfly/quickstarts';
 
-export const QuickStartCatalog = ({ ...props }) => <PfQuickStartCatalogPage {...props} />;
+const QuickStartCatalog = ({ ...props }) => <PfQuickStartCatalogPage {...props} />;
+
+export default QuickStartCatalog;
